requirements.txt
Django==1.7.7
django-admin-tools==0.5.2
django-configurations==0.8
django-debug-toolbar==1.3.0
django-filter==0.9.2
django-jquery==1.9.1
djangorestframework==3.1.1
grin==1.2.1
Markdown==2.6.1
pep8==1.6.2
pyreadline==2.0
sqlparse==0.1.14
